+++ gantry/util.py
@@ -13,6 +13,7 @@ import requests
 import rich
 from beaker import (
     Beaker,
+    BeakerCancelationCode,
     BeakerDataset,
     BeakerDatasetFileAlgorithmType,
     BeakerJob,
@@ -30,6 +31,7 @@ from git.cmd import Git
 from git.repo import Repo
 from rich import print, prompt
 from rich.console import Console
+from rich.status import Status
 
 from . import constants
 from .exceptions import *
@@ -121,6 +123,43 @@ def get_latest_workload(
         return None
 
 
+def _wait_for_job_to_start(
+    *,
+    beaker: Beaker,
+    job: BeakerJob,
+    status: Status,
+    start_time: float,
+    timeout: int = 0,
+    show_logs: bool = True,
+) -> BeakerJob:
+    # Pull events until job is running (or fails)...
+    events = set()
+    while not (job.status.HasField("finalized") or (show_logs and job.status.HasField("started"))):
+        if timeout > 0 and (time.monotonic() - start_time) > timeout:
+            raise BeakerJobTimeoutError(f"Timed out while waiting for job '{job.id}' to finish")
+
+        for event in beaker.job.list_summarized_events(
+            job, sort_order=BeakerSortOrder.descending, sort_field="latest_occurrence"
+        ):
+            event_hashable = (event.latest_occurrence.ToSeconds(), event.latest_message)
+            if event_hashable not in events:
+                status.update(f"[i]{event.latest_message}[/]")
+                events.add(event_hashable)
+                time.sleep(0.5)
+
+        time.sleep(0.5)
+        job = beaker.job.get(job.id)
+
+    return job
+
+
+def _job_preempted(job: BeakerJob) -> bool:
+    return job.status.status == BeakerWorkloadStatus.canceled and job.status.canceled_code in (
+        BeakerCancelationCode.system_preemption,
+        BeakerCancelationCode.user_preemption,
+    )
+
+
 def follow_workload(
     beaker: Beaker,
     workload: BeakerWorkload,
@@ -130,53 +169,58 @@ def follow_workload(
 ) -> BeakerJob:
     console = rich.get_console()
     start_time = time.monotonic()
+    preempted_job_ids = set()
 
-    with console.status("[i]waiting...[/]", spinner="point", speed=0.8) as status:
-        # Wait for job to start...
-        while (job := beaker.workload.get_latest_job(workload)) is None:
-            time.sleep(1.0)
+    while True:
+        with console.status("[i]waiting...[/]", spinner="point", speed=0.8) as status:
+            # Wait for job to be created...
+            job: BeakerJob | None = None
+            while job is None:
+                if (
+                    j := beaker.workload.get_latest_job(workload)
+                ) is not None and j.id not in preempted_job_ids:
+                    job = j
+                else:
+                    time.sleep(1.0)
 
-        # Pull events until job is running (or fails)...
-        events = set()
+            # Wait for job to start...
+            job = _wait_for_job_to_start(
+                beaker=beaker,
+                job=job,
+                status=status,
+                start_time=start_time,
+                timeout=timeout,
+                show_logs=show_logs,
+            )
+
+        # Stream logs...
+        if show_logs and job.status.HasField("started"):
+            print()
+            rich.get_console().rule("Logs")
+
+            for job_log in beaker.job.logs(job, tail_lines=10 if tail else None, follow=True):
+                console.print(job_log.message.decode(), highlight=False, markup=False)
+                if timeout > 0 and (time.monotonic() - start_time) > timeout:
+                    raise BeakerJobTimeoutError(
+                        f"Timed out while waiting for job '{job.id}' to finish"
+                    )
+
+            print()
+            rich.get_console().rule("End logs")
+            print()
+
+        # Wait for job to finalize...
         while not job.status.HasField("finalized"):
-            if timeout > 0 and (time.monotonic() - start_time) > timeout:
-                raise BeakerJobTimeoutError(f"Timed out while waiting for job '{job.id}' to finish")
-
+            time.sleep(0.5)
             job = beaker.job.get(job.id)
 
-            for event in beaker.job.list_summarized_events(
-                job, sort_order=BeakerSortOrder.descending, sort_field="latest_occurrence"
-            ):
-                event_hashable = (event.latest_occurrence.ToSeconds(), event.latest_message)
-                if event_hashable not in events:
-                    status.update(f"[i]{event.latest_message}[/]")
-                    events.add(event_hashable)
-                    if show_logs and event.status.lower() == "started":
-                        break
-            else:
-                time.sleep(1.0)
-                continue
+        # If job was preempted, we start over...
+        if _job_preempted(job):
+            print(f"[yellow]Job '{job.id}' preempted.[/] ")
+            preempted_job_ids.add(job.id)
+            continue
 
-            break
-
-    if show_logs:
-        print()
-        rich.get_console().rule("Logs")
-        time.sleep(2.0)  # wait a moment to make sure logs are available before experiment finishes
-        for job_log in beaker.job.logs(job, tail_lines=10 if tail else None, follow=True):
-            console.print(job_log.message.decode(), highlight=False, markup=False)
-            if timeout > 0 and (time.monotonic() - start_time) > timeout:
-                raise BeakerJobTimeoutError(f"Timed out while waiting for job '{job.id}' to finish")
-        print()
-        rich.get_console().rule("End logs")
-        print()
-
-    # Wait for job to finalize...
-    while not job.status.HasField("finalized"):
-        time.sleep(0.5)
-        job = beaker.job.get(job.id)
-
-    return job
+        return job
 
 
 def display_logs(beaker: Beaker, job: BeakerJob, tail_lines: Optional[int] = None) -> BeakerJob: