+++ CHANGELOG.md
@@ -13,7 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Fixed filtering with `--all` in `gantry list` command.
+- (bug) Fixed filtering with `--all` in `gantry list` command.
+- (performance) Made `gantry list` much faster by using a thread pool to query for task status.
 
 ## [v2.1.1](https://github.com/allenai/beaker-gantry/releases/tag/v2.1.1) - 2025-05-09