+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added `gantry open` command for opening Beaker resources in your browser.
 
+### Changed
+
+- Gantry will now continue to follow experiments after their main job gets preempted.
+
 ### Fixed
 
 - (bug) Fixed filtering with `--all` in `gantry list` command.